Convo V2
Convo V2 Analyser is able to plot 720 point data of conversation between client and company with the ability to detect delays and Critical points in conversation depending on conversation flow. The program is also able to make the different types of charts from the Datapoints. This is the main data plotter program and works in conjunction with the Data  Displayer program.